Move MemberFilter into Domain
